Update 1042_Lab1B_group12.py (#16)
See above commentary.. wrote it in the wrong spot I guess (wasn't sure it would give me an option for a comment at the time of the merge -- it DOES). Note: Still need to check "resolve conflicts" functionality merging master onto slave branch if that's the first PR between the two (see if it works then) -- and need to look up what "rebase and merge" does -- right now, it's looking like "Squash and merge" is the option we want to go for.